# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TotHub Daycare Management API
  description: |
    Enterprise-grade childcare management platform API for attendance tracking, 
    staff scheduling, compliance monitoring, and parent communication.
    
    ## Key Features
    - Multi-tenant architecture with role-based access control
    - Real-time attendance and ratio monitoring
    - Multi-state compliance rules engine
    - Staff scheduling and payroll management
    - Parent portal and communication tools
    - Biometric authentication support
    - QuickBooks integration
    
    ## Authentication
    All API calls require a valid JWT token in the Authorization header:
    `Authorization: Bearer <token>`
    
    ## Rate Limiting
    - Standard endpoints: 1000 requests/hour
    - Authentication endpoints: 100 requests/hour
    - File uploads: 100 requests/hour
  
  version: 1.0.0
  contact:
    name: TotHub Development Team
    email: dev@tothub.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.tothub.com/v1
    description: Production server
  - url: https://staging-api.tothub.com/v1
    description: Staging server
  - url: http://localhost:5000/v1
    description: Local development
security:
  - bearerAuth: []
paths:
...
  /children:
    get:
      summary: List children
      description: Retrieve paginated list of children with optional filtering
      tags: [Children]
      parameters:
        - name: page
          in: query
          description: Page number (1-based)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: room
          in: query
          description: Filter by room ID
          schema:
            type: string
        - name: status
          in: query
          description: Filter by enrollment status
          schema:
            type: string
            enum: [enrolled, unenrolled, aged_out]
      responses:
        '200':
          description: List of children
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Child'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '403':
          description: Insufficient permissions

    post:
      summary: Create new child
      description: Add a new child to the system
      tags: [Children]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChildCreate'
      responses:
        '201':
          description: Child created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Child'
        '400':
          description: Validation error
        '403':
          description: Insufficient permissions
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Child:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
          format: date
        parentEmail:
          type: string
          format: email
        parentPhone:
          type: string
        emergencyContact:
          type: string
        medicalNotes:
          type: string
        enrollmentDate:
          type: string
          format: date
        isActive:
          type: boolean
        roomId:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
    ChildCreate:
      type: object
      required: [firstName, lastName, dateOfBirth, parentEmail]
      properties:
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        dateOfBirth:
          type: string
          format: date
        parentEmail:
          type: string
          format: email
        parentPhone:
          type: string
        emergencyContact:
          type: string
        medicalNotes:
          type: string
        roomId:
          type: string
          format: uuid
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 1
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
...